# openapi.yaml
openapi: 3.0.3
info:
  title: ias_stt — Trap Management API
  version: 1.0.0
  description: |
    Trap management service with deployment tracking, location
    history, and immutable deployment actions with required pictures. Telemetry is ingested from the
    ChirpStack HTTP integration via `POST /api/telemetry/ingest`
    (an optional MQTT listener remains available). Application users obtain
    JWT access and refresh tokens from `/auth/login` and send access tokens in
    the `Authorization` header. A legacy API key remains available for service
    integrations.
servers:
  - url: http://localhost:8080
    description: Local development
security:
  - bearerAuth: []
...
paths:
...
  /auth/login:
    post:
      tags: [Auth]
      summary: Authenticate a LOCAL or LDAP-provisioned user
      operationId: login
      description: |
        LDAP credentials are accepted only over HTTPS. The backend selects the
        provider from the provisioned local application account. LDAP failures
        never fall back to LOCAL password verification. Both providers receive
        the same application JWT response.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [username, password]
              properties:
                username: { type: string }
                password: { type: string, format: password }
      responses:
        200:
          description: Access and refresh tokens
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        400:
          $ref: '#/components/responses/400'
        401:
          $ref: '#/components/responses/401'
        403:
          $ref: '#/components/responses/403'
        503:
          $ref: '#/components/responses/503'
...
  /auth/me:
    get:
      tags: [Auth]
      summary: Get the current authenticated user
      operationId: currentUser
      responses:
        200:
          description: Current user and permissions
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthMeResponse'
        401:
          $ref: '#/components/responses/401'
  /auth/refresh:
    post:
      tags: [Auth]
      summary: Exchange a refresh token for an access token
      operationId: refreshToken
      responses:
        200:
          description: New access token
          content:
            application/json:
              schema:
                type: object
                properties:
                  access_token: { type: string }
                  token_type: { type: string, example: Bearer }
                  expires_in: { type: integer, example: 900 }
        401:
          $ref: '#/components/responses/401'
        503:
          $ref: '#/components/responses/503'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: |
        Application users pass the JWT access token as
        `Authorization: Bearer <access-token>`. Legacy service integrations
        may pass the configured API key in the same header.

  # -- Reusable parameters ---------------------------------------------------
...
  schemas:
    Error:
      type: object
      properties:
        error:
          oneOf:
            - type: string
            - type: object
              properties:
                code: { type: string }
                message: { type: string }
      example:
        error:
          code: AUTHENTICATION_REQUIRED
          message: Authentication is required.
    AuthUser:
      type: object
      properties:
        id: { type: integer }
        username: { type: string }
        display_name: { type: string, nullable: true }
        email: { type: string, nullable: true }
        role: { type: string, enum: [administrator, field_operator, read_only] }
        auth_provider: { type: string, enum: [LOCAL, LDAP] }
        permissions:
          type: array
          items: { type: string }
...
    LoginResponse:
      type: object
      properties:
        authenticated: { type: boolean }
        user: { $ref: '#/components/schemas/AuthUser' }
        access_token: { type: string }
        refresh_token: { type: string }
        token_type: { type: string, example: Bearer }
        expires_in: { type: integer, example: 900 }
        refresh_expires_in: { type: integer, example: 604800 }
        expires_at: { type: string, format: date-time }
        refresh_expires_at: { type: string, format: date-time }
    AuthMeResponse:
      type: object
      properties:
        authenticated: { type: boolean }
        user: { $ref: '#/components/schemas/AuthUser' }
        expires_at: { type: string, format: date-time }
...
  responses:
    400:
      description: Bad request — missing or invalid fields
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          examples:
            missingField:
              summary: Missing required field
              value: { error: "Missing required field(s): status" }
            invalidType:
              summary: Invalid limit/offset
              value: { error: "'limit' and 'offset' must be integers" }
    401:
      description: Missing or invalid JWT or service credential
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error:
              code: AUTHENTICATION_REQUIRED
              message: Authentication is required.
    403:
      description: Authenticated user lacks the required permission
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error:
              code: FORBIDDEN
              message: You do not have permission to perform this action.
...
    503:
      description: The configured LDAP authentication provider is unavailable or misconfigured
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error:
              code: AUTH_PROVIDER_UNAVAILABLE
              message: The authentication service is temporarily unavailable.